Renamed .d.ts to align with terraformer.js name.
